// src/features/MasterModal/models/formInputs.ts
interface FormInputProps {
  placeholder: string;
  value: string;
  fieldName: string;
}

export const formInputs: Array<FormInputProps> = [
  // {
  //   placeholder: "Familiya",
  //   value: lastName,
  //   fieldName: "lastName",
  // },
  // {
  //   placeholder: "Ism",
  //   value: firstName,
  //   fieldName: "firstName",
  // },
  // {
  //   placeholder: "Ism sharif",
  //   value: middleName,
  //   fieldName: "middleName",
  // },
];